price now accuate when shown, empty cart can be seen when clicking on new cart symbol in nav bar, Registring and Login added, Registring: data is sendo to database, login: not working yet, pages enhanced, bugs fixed

+++ DataLibrary/Logic/ProductProcessor.cs
@@ -11,7 +11,7 @@ namespace DataLibrary.Logic
 {
     public static class ProductProcessor
     {
-        public static int CreateProduct(string titel, float price, string photo, string content, string author)
+        public static int CreateProduct(string titel, decimal price, string photo, string content, string author)
         {
             Product data = new Product
             {
